api/flaskapp/requirements.txt
Flask
gunicorn
python-dotenv
requests
web3